medical-ai-chatbot-template/requirements.txt
# Flask and Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.0

# Data Processing - Use latest compatible versions
pandas>=2.0.0
numpy>=1.21.0

# AI and ML
requests>=2.25.0

# File Processing
PyPDF2>=3.0.0
Pillow>=9.0.0
pytesseract>=0.3.8
openpyxl>=3.0.0
xlrd>=2.0.0

# Configuration
PyYAML>=6.0.0
python-dotenv>=0.19.0


# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# AI and HTTP Requests
requests==2.31.0

# File Processing
PyPDF2==3.0.1

# Configuration Management
PyYAML==6.0.1
python-dotenv==1.0.0

# Utilities
uuid==1.30

# Optional: For enhanced features
# Pillow==10.1.0          # Image processing
# pandas==2.1.4           # Advanced data analysis
# pytesseract==0.3.10     # OCR for images
# openpyxl==3.1.2         # Excel file support